moved MSLevel declaration
