+++ src/main/java/net/sf/mzmine/parameters/parametertypes/ProcessingParameter.java
@@ -2,9 +2,6 @@ package net.sf.mzmine.parameters.parametertypes;
 
 import java.util.Collection;
 import org.w3c.dom.Element;
-import net.sf.mzmine.modules.visualization.spectra.simplespectra.datapointprocessing.DataPointProcessingManager;
-import net.sf.mzmine.modules.visualization.spectra.simplespectra.datapointprocessing.DataPointProcessingManager.MSLevel;
-import net.sf.mzmine.modules.visualization.spectra.simplespectra.datapointprocessing.DataPointProcessingQueue;
 import net.sf.mzmine.modules.visualization.spectra.simplespectra.datapointprocessing.datamodel.DPPParameterValueWrapper;
 import net.sf.mzmine.parameters.UserParameter;
